Untangle ScannerEngineInput and ScannerEngineInputGenerator
