+++ Tests/SonarScanner.MSBuild.Shim.Test/ScannerEngineInputGeneratorTest.TryWriteProperties.cs
@@ -18,6 +18,8 @@
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
 
+using Newtonsoft.Json;
+
 namespace SonarScanner.MSBuild.Shim.Test;
 
 public partial class ScannerEngineInputGeneratorTest
@@ -25,7 +27,7 @@ public partial class ScannerEngineInputGeneratorTest
     [TestMethod]
     public void TryWriteProperties_WhenThereIsNoCommonPath_LogsError()
     {
-        var outPath = Path.Combine(TestContext.TestRunDirectory!, ".sonarqube", "out");
+        var outPath = Path.Combine(TestContext.TestRunDirectory, ".sonarqube", "out");
         Directory.CreateDirectory(outPath);
         var fileToAnalyzePath = TestUtils.CreateEmptyFile(TestContext.TestRunDirectory, "file.cs");
         var filesToAnalyzePath = TestUtils.CreateFile(TestContext.TestRunDirectory, TestUtils.FilesToAnalyze, fileToAnalyzePath);
@@ -133,6 +135,163 @@ public partial class ScannerEngineInputGeneratorTest
         new ScannerEngineInputReader(engineInput.ToString()).AssertProperty("sonar.host.url", "http://localhost:9000");
     }
 
+    [TestMethod]
+    public void WriteAnalyzerOutputPaths_ForUnexpectedLanguage_DoNotWritesOutPaths()
+    {
+        var input = new ScannerEngineInput(new());
+        ScannerEngineInputGenerator.WriteAnalyzerOutputPaths(input, CreateProjectDataWithPaths("unexpected", analyzerOutPaths: [@"c:\dir1\dir2"]));
+
+        input.ToString().Should().BeIgnoringLineEndings("""
+            {
+              "scannerProperties": [
+                {
+                  "key": "sonar.modules",
+                  "value": ""
+                }
+              ]
+            }
+            """);
+    }
+
+    [TestMethod]
+    [DataRow(ProjectLanguages.CSharp, "sonar.cs.analyzer.projectOutPaths")]
+    [DataRow(ProjectLanguages.VisualBasic, "sonar.vbnet.analyzer.projectOutPaths")]
+    public void WriteAnalyzerOutputPaths_WritesEncodedPaths(string language, string expectedPropertyKey)
+    {
+        var input = new ScannerEngineInput(new());
+        ScannerEngineInputGenerator.WriteAnalyzerOutputPaths(input, CreateProjectDataWithPaths(
+            language,
+            analyzerOutPaths: [Path.Combine(TestUtils.DriveRoot(), "dir1", "first"), Path.Combine(TestUtils.DriveRoot(), "dir1", "second")]));
+
+        input.ToString().Should().BeIgnoringLineEndings($$"""
+            {
+              "scannerProperties": [
+                {
+                  "key": "sonar.modules",
+                  "value": ""
+                },
+                {
+                  "key": "5762C17D-1DDF-4C77-86AC-E2B4940926A9.{{expectedPropertyKey}}",
+                  "value": {{JsonConvert.ToString(Path.Combine(TestUtils.DriveRoot(), "dir1", "first") + "," + Path.Combine(TestUtils.DriveRoot(), "dir1", "second"))}}
+                }
+              ]
+            }
+            """);
+    }
+
+    [TestMethod]
+    public void WriteRoslynReportPaths_ForUnexpectedLanguage_DoNotWritesOutPaths()
+    {
+        var input = new ScannerEngineInput(new());
+        ScannerEngineInputGenerator.WriteRoslynReportPaths(input, CreateProjectDataWithPaths("unexpected"));
+
+        input.ToString().Should().BeIgnoringLineEndings("""
+            {
+              "scannerProperties": [
+                {
+                  "key": "sonar.modules",
+                  "value": ""
+                }
+              ]
+            }
+            """);
+    }
+
+    [TestMethod]
+    [DataRow(ProjectLanguages.CSharp, "sonar.cs.roslyn.reportFilePaths")]
+    [DataRow(ProjectLanguages.VisualBasic, "sonar.vbnet.roslyn.reportFilePaths")]
+    public void WriteRoslynReportPaths_WritesEncodedPaths(string language, string expectedPropertyKey)
+    {
+        var input = new ScannerEngineInput(new());
+        ScannerEngineInputGenerator.WriteRoslynReportPaths(input, CreateProjectDataWithPaths(language, roslynOutPaths: [
+            Path.Combine(TestUtils.DriveRoot(), "dir1", "first"),
+            Path.Combine(TestUtils.DriveRoot(), "dir1", "second")]));
+
+        input.ToString().Should().BeIgnoringLineEndings($$"""
+            {
+              "scannerProperties": [
+                {
+                  "key": "sonar.modules",
+                  "value": ""
+                },
+                {
+                  "key": "5762C17D-1DDF-4C77-86AC-E2B4940926A9.{{expectedPropertyKey}}",
+                  "value": {{JsonConvert.ToString(Path.Combine(TestUtils.DriveRoot(), "dir1", "first") + "," + Path.Combine(TestUtils.DriveRoot(), "dir1", "second"))}}
+                }
+              ]
+            }
+            """);
+    }
+
+    [TestMethod]
+    public void Telemetry_ForUnexpectedLanguage_DoNotWritePaths()
+    {
+        var input = new ScannerEngineInput(new());
+        ScannerEngineInputGenerator.WriteTelemetryPaths(input, CreateProjectDataWithPaths("unexpected", telemetryPaths: [@"c:\dir1\dir2\Telemetry.json"]));
+
+        input.ToString().Should().BeIgnoringLineEndings("""
+            {
+              "scannerProperties": [
+                {
+                  "key": "sonar.modules",
+                  "value": ""
+                }
+              ]
+            }
+            """);
+    }
+
+    [TestMethod]
+    [DataRow(ProjectLanguages.CSharp, "sonar.cs.scanner.telemetry")]
+    [DataRow(ProjectLanguages.VisualBasic, "sonar.vbnet.scanner.telemetry")]
+    public void Telemetry_WritesEncodedPaths(string language, string expectedPropertyKey)
+    {
+        var input = new ScannerEngineInput(new());
+        ScannerEngineInputGenerator.WriteTelemetryPaths(input, CreateProjectDataWithPaths(language, telemetryPaths: [
+            Path.Combine(TestUtils.DriveRoot(), "dir1", "first", "Telemetry.json"),
+            Path.Combine(TestUtils.DriveRoot(), "dir1", "second", "Telemetry.json")]));
+
+        input.ToString().Should().BeIgnoringLineEndings($$"""
+            {
+              "scannerProperties": [
+                {
+                  "key": "sonar.modules",
+                  "value": ""
+                },
+                {
+                  "key": "5762C17D-1DDF-4C77-86AC-E2B4940926A9.{{expectedPropertyKey}}",
+                  "value": {{JsonConvert.ToString(Path.Combine(TestUtils.DriveRoot(), "dir1", "first", "Telemetry.json") + "," + Path.Combine(TestUtils.DriveRoot(), "dir1", "second", "Telemetry.json"))}}
+                }
+              ]
+            }
+            """);
+    }
+
+    [TestMethod]
+    public void TryWriteProperties_ProjectAnalysisSettings_Propagated()
+    {
+        var config = new AnalysisConfig { SonarOutputDir = TestContext.TestRunDirectory };
+        var input = new ScannerEngineInput(config);
+        var project = CreateProjectDataWithPaths(ProjectLanguages.CSharp);
+        project.Status = ProjectInfoValidity.Valid;
+        project.Project.FullPath = TestUtils.CreateEmptyFile(TestContext.TestRunDirectory, "Project.proj");
+        project.Project.AnalysisSettings =
+        [
+            new("my.setting1", "setting1"),
+            new("my.setting2", "setting 2 with spaces"),
+            new("my.setting.3", @"c:\dir1\dir2\foo.txt")
+        ];
+        project.ReferencedFiles.Add(new(TestUtils.CreateEmptyFile(TestContext.TestRunDirectory, "File.cs")));
+        var sut = new ScannerEngineInputGenerator(config, logger);
+        sut.TryWriteProperties(config.ToAnalysisProperties(logger), [project], new PropertiesWriter(config), input);
+
+        logger.AssertNoErrorsLogged();
+        var reader = new ScannerEngineInputReader(input.ToString());
+        reader.AssertProperty("5762C17D-1DDF-4C77-86AC-E2B4940926A9.my.setting1", "setting1");
+        reader.AssertProperty("5762C17D-1DDF-4C77-86AC-E2B4940926A9.my.setting2", "setting 2 with spaces");
+        reader.AssertProperty("5762C17D-1DDF-4C77-86AC-E2B4940926A9.my.setting.3", @"c:\dir1\dir2\foo.txt");
+    }
+
     private void TryWriteProperties_HostUrl_Execute(AnalysisConfig config, PropertiesWriter legacyWriter, ScannerEngineInput engineInput)
     {
         Directory.CreateDirectory(config.SonarOutputDir);
@@ -155,5 +314,26 @@ public partial class ScannerEngineInputGeneratorTest
             legacyWriter,
             engineInput)
             .Should().BeTrue();
+    }
+
+    private static ProjectData CreateProjectDataWithPaths(string language, string[] analyzerOutPaths = null, string[] roslynOutPaths = null, string[] telemetryPaths = null)
+    {
+        analyzerOutPaths ??= [];
+        roslynOutPaths ??= [];
+        telemetryPaths ??= [];
+        var projectData = new[] { new ProjectInfo { ProjectGuid = new("5762C17D-1DDF-4C77-86AC-E2B4940926A9"), ProjectLanguage = language } }.ToProjectData(true, Substitute.For<ILogger>()).Single();
+        foreach (var path in analyzerOutPaths)
+        {
+            projectData.AnalyzerOutPaths.Add(new FileInfo(path));
+        }
+        foreach (var path in roslynOutPaths)
+        {
+            projectData.RoslynReportFilePaths.Add(new FileInfo(path));
+        }
+        foreach (var path in telemetryPaths)
+        {
+            projectData.TelemetryPaths.Add(new FileInfo(path));
+        }
+        return projectData;
     }
 }
